api/pyproject.toml
[tool.poetry]
name = "portfolio-analytics-api"
version = "1.0.0"
description = "Backend API for portfolio analytics platform"
authors = ["Portfolio Analytics Team"]
readme = "README.md"
packages = [{include = "api"}, {include = "domain"}, {include = "adapters"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
psycopg = {extras = ["binary", "pool"], version = "^3.1.0"}
duckdb = "^0.10.0"
pydantic = "^2.5.0"
pyyaml = "^6.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
ruff = "^0.2.0"
mypy = "^1.8.0"
types-pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["duckdb", "psycopg_pool"]
ignore_missing_imports = true
